refatorar processDatabase: substituir executeQuery por runQuery com tratamento de erros e validações específicas
