+++ src/utils/processDatabase.js
@@ -1,32 +1,3 @@
-/**
- * @file processDatabase.js
- * @description Módulo para inicializar e configurar um banco de dados MySQL e suas tabelas.
- *
- * Este módulo valida as variáveis de ambiente essenciais e exporta um
- * objeto de configuração MySQL (databaseConfig) juntamente com uma função assíncrona (initDatabase)
- * que:
- *  - Cria o banco de dados especificado (ou um padrão caso não seja informado),
- *  - Estabelece uma conexão segura com o banco de dados.
- *
- * Variáveis de Ambiente:
- *   - MYSQL_LOGIN_USER (string): Obrigatório. O usuário do MySQL.
- *   - MYSQL_LOGIN_PASSWORD (string): Obrigatório. A senha do usuário do MySQL.
- *   - MYSQL_HOST (string): Opcional. O host do MySQL (padrão: "localhost").
- *   - MYSQL_DATABASE (string): Opcional. O nome do banco de dados (padrão: "cat").
- *
- * Exporta:
- * @module processDatabase
- * @property {Object} databaseConfig - Objeto de configuração para criar uma conexão MySQL.
- * @property {Function} initDatabase - Função assíncrona que inicializa o banco de dados.
- *
- * @throws {Error} Lança um erro se qualquer variável de ambiente necessária estiver faltando,
- * ou se ocorrer qualquer erro na conexão do banco ou criação das tabelas.
- *
- * @async
- * @function initDatabase
- * @returns {Promise<Connection>} Uma promessa que resolve para uma conexão MySQL segura com o banco de dados especificado.
- */
-
 const mysql = require("mysql2/promise");
 const logger = require("./logger");
 require("dotenv").config();
@@ -48,8 +19,6 @@ const databaseConfig = {
   connectionLimit: 10,
   charset: "utf8mb4",
 };
-
-module.exports = databaseConfig;
 
 let connection; // Variável para armazenar a conexão compartilhada
 
@@ -79,23 +48,58 @@ async function initDatabase() {
   return connection;
 }
 
-/* Executa uma query no banco de dados */
-async function executeQuery(query, params) {
+/* 
+Executa uma query com tratamento de erros e validações específicas por tipo de operação.
+*/
+async function runQuery(query, params = []) {
   try {
     if (!connection) {
       throw new Error("Conexão com o banco de dados não inicializada.");
     }
-    const [results] = await connection.execute(query, params);
-    return results;
-  } catch (error) {
-    logger.error(`Erro ao executar query: ${error.stack}`);
-    throw error;
+
+    const [result] = await connection.execute(query, params);
+
+    // Identifica o tipo de query
+    const queryType = query.trim().split(" ")[0].toUpperCase();
+    const isIgnoreQuery = query.toUpperCase().includes("INSERT IGNORE");
+
+    // Validações e retornos específicos por tipo
+    switch (queryType) {
+      case "SELECT":
+        if (!result || result.length === 0) {
+          logger.debug(`⚠️ Nenhum resultado encontrado para a consulta`);
+          return [];
+        }
+        return result;
+
+      case "INSERT":
+        if (!result.affectedRows && !isIgnoreQuery) {
+          throw new Error("Nenhuma linha foi inserida");
+        }
+        return {
+          insertId: result.insertId,
+          affectedRows: result.affectedRows,
+        };
+
+      case "UPDATE":
+      case "DELETE":
+        return {
+          affectedRows: result.affectedRows,
+          changedRows: result.changedRows,
+        };
+
+      default:
+        return result;
+    }
+  } catch (err) {
+    logger.error(`❌ Erro ao executar query:\n→ Query: ${query}\n→ Parâmetros: ${JSON.stringify(params)}\n→ Detalhes: ${err.message}`);
+    throw new Error(`Erro na execução da consulta: ${err.message}`);
   }
 }
 
 module.exports = {
   databaseConfig,
   initDatabase,
-  connection, // Exporta a conexão compartilhada
-  executeQuery, // Exporta a função para executar queries
+  connection,
+  runQuery, // Adicionando runQuery às exportações
 };
